// openapi/models/party-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Skrib
 * Skrib
 *
 * The version of the OpenAPI document: 2.11.7
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface PartyResponse
 */
export interface PartyResponse {
    /**
     * 
     * @type {string}
     * @memberof PartyResponse
     */
    'email': string;
    /**
     * 
     * @type {string}
     * @memberof PartyResponse
     */
    'stipulationstatus': string;
    /**
     * 
     * @type {number}
     * @memberof PartyResponse
     */
    'stipulationupdatedby': number;
    /**
     * 
     * @type {string}
     * @memberof PartyResponse
     */
    'stipulationupdatedat': string;
    /**
     * 
     * @type {object}
     * @memberof PartyResponse
     */
    'stipulationupdatedbyinfo': object;
    /**
     * 
     * @type {number}
     * @memberof PartyResponse
     */
    'id': number;
    /**
     * 
     * @type {number}
     * @memberof PartyResponse
     */
    'eventid': number;
    /**
     * 
     * @type {string}
     * @memberof PartyResponse
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof PartyResponse
     */
    'firstname'?: string;
    /**
     * 
     * @type {string}
     * @memberof PartyResponse
     */
    'lastname'?: string;
    /**
     * 
     * @type {string}
     * @memberof PartyResponse
     */
    'partytype': string;
    /**
     * 
     * @type {number}
     * @memberof PartyResponse
     */
    'tenantid': number;
    /**
     * 
     * @type {string}
     * @memberof PartyResponse
     */
    'invitationcode': string;
}