// openapi/models/tenant-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Skrib
 * Skrib
 *
 * The version of the OpenAPI document: 2.11.7
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface TenantResponse
 */
export interface TenantResponse {
    /**
     * 
     * @type {number}
     * @memberof TenantResponse
     */
    'id': number;
    /**
     * 
     * @type {string}
     * @memberof TenantResponse
     */
    'serviceAccount': string;
    /**
     * 
     * @type {string}
     * @memberof TenantResponse
     */
    'bucket': string;
    /**
     * 
     * @type {string}
     * @memberof TenantResponse
     */
    'tenantkey': string;
    /**
     * 
     * @type {string}
     * @memberof TenantResponse
     */
    'email': string;
    /**
     * 
     * @type {string}
     * @memberof TenantResponse
     */
    'firstname': string;
    /**
     * 
     * @type {string}
     * @memberof TenantResponse
     */
    'lastname': string;
    /**
     * 
     * @type {string}
     * @memberof TenantResponse
     */
    'company': string;
    /**
     * 
     * @type {string}
     * @memberof TenantResponse
     */
    'phone'?: string;
    /**
     * 
     * @type {string}
     * @memberof TenantResponse
     */
    'country': string;
    /**
     * 
     * @type {string}
     * @memberof TenantResponse
     */
    'state'?: string;
    /**
     * 
     * @type {string}
     * @memberof TenantResponse
     */
    'invitationcode'?: string;
}